+++ metadata_backend/api/operators.py
@@ -734,13 +734,13 @@ class SubmissionOperator:
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
 
-    async def check_object_in_submission(self, collection: str, accession_id: str) -> Tuple[bool, str, bool]:
+    async def check_object_in_submission(self, collection: str, accession_id: str) -> Tuple[str, bool]:
         """Check a object/draft is in a submission.
 
         :param collection: collection it belongs to, it would be used as path
         :param accession_id: document by accession_id
         :raises: HTTPUnprocessableEntity if error occurs during the process and object in more than 1 submission
-        :returns: Tuple with True for the check, submission id and if published or not
+        :returns: Tuple with submission id if object is in submission and bool if published or not
         """
         try:
             submission_path = "drafts" if collection.startswith("draft") else "metadataObjects"
@@ -756,7 +756,7 @@ class SubmissionOperator:
 
         if len(submission_check) == 0:
             LOG.info("Doc with accession ID: %r belongs to no submission something is off", accession_id)
-            return False, "", False
+            return "", False
 
         if len(submission_check) > 1:
             reason = f"The {accession_id} is in more than 1 submission."
@@ -765,7 +765,7 @@ class SubmissionOperator:
 
         submission_id = submission_check[0]["submissionId"]
         LOG.info("Found doc with accession ID: %r in submission: %r.", accession_id, submission_id)
-        return True, submission_id, submission_check[0]["published"]
+        return submission_id, submission_check[0]["published"]
 
     async def get_collection_objects(self, submission_id: str, collection: str) -> List:
         """List objects ids per collection.
@@ -939,7 +939,7 @@ class SubmissionOperator:
     async def delete_submission(self, submission_id: str) -> Union[str, None]:
         """Delete object submission from database.
 
-        :param submission_id: ID of the submission to delete.
+        :param submission_id: ID of the submission to delete
         :raises: HTTPBadRequest if deleting was not successful
         :returns: ID of the submission deleted from database
         """
@@ -960,6 +960,7 @@ class SubmissionOperator:
     async def check_submission_exists(self, submission_id: str) -> None:
         """Check the existance of a submission by its id in the database.
 
+        :param submission_id: ID of the submission to check
         :raises: HTTPNotFound if submission does not exist
         """
         exists = await self.db_service.exists("submission", submission_id)
@@ -968,16 +969,18 @@ class SubmissionOperator:
             LOG.error(reason)
             raise web.HTTPNotFound(reason=reason)
 
-    async def check_submission_published(self, submission_id: str) -> None:
-        """Check whether a submission in the database is published.
+    async def check_submission_published(self, submission_id: str, method: str) -> None:
+        """Check the published status of a submission by its id in the database.
 
-        :raises: HTTPNotFound if submission does not exist
+        :param submission_id: ID of the submission to check
+        :param method: Name of HTTP method used when this check is executed
+        :raises: HTTPMethodNotAllowed if submission is not published
         """
         published = await self.db_service.published_submission(submission_id)
         if published:
             reason = f"Submission with ID: '{submission_id}' is already published and cannot be modified or deleted."
             LOG.error(reason)
-            raise web.HTTPUnauthorized(reason=reason)
+            raise web.HTTPMethodNotAllowed(method=method, allowed_methods=["GET", "HEAD"], reason=reason)
 
     def _generate_submission_id(self) -> str:
         """Generate random submission id.